Have some CI fun: Caching, test all crates
The idea is this: Since the code from the guides is now checked in as
crates in `examples/`, we can call `cargo test --all` in the one CI
build and only use the waltz build step to assert that they are up to
date.

While we're at it, let's also turn on caching.

cf .#63