Deprecate main macro and structopt re-export
This is quite a big change! Quicli is now fully invested in the Rust
2018 edition, and with it lost some of the most fragile features.

First off, the changes to the module system made it quite unwieldy to
re-export the structopt trait as well as the derive macro. The latter
always assumed that a `structopt` module existed at the root level,
which we cheated by exporting a fake `structopt` module in the prelude
and told people to have `use quicli::prelude::*;` at the root level.
Super hacky, and now deleted.

Next up, the `main!` macro is gone. The guides have been updated, and
the necessary `Ok(())` lines have been added. Thanks to the now-
available `?` in main feature, and our export of the `CliResult` type,
it seemed like a good idea to switch away from the macro. This of course
comes at the cost of having to tell people to call `Cli::from_args();`
in their code, and, to set up logging, even add a line like
`args.verbosity.setup_env_logger("crate-name")?;`.

I think this is worth it since we are now doing much less magical
things.